feat: add wecom and wecomApp test
